+++ src/main/java/com/springboot/restapiwebservices/controller/PurchaseOrderController.java
@@ -3,59 +3,84 @@ package com.springboot.restapiwebservices.controller;
 import com.springboot.restapiwebservices.model.ProjectContactModel;
 import com.springboot.restapiwebservices.model.PurchaseOrderModel;
 import com.springboot.restapiwebservices.repository.PurchaseOrderRepo;
+import com.springboot.restapiwebservices.service.PurchaseOrderService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
+@RequestMapping("/rest/PurchaseOrder")
 public class PurchaseOrderController {
 
     @Autowired
     PurchaseOrderRepo purchaseOrderRepo;
+    @Autowired
+    PurchaseOrderService purchaseOrderService;
 
-    @PostMapping("/saveOrder")
-    public String insertOrder(@RequestBody PurchaseOrderModel purchaseOrderModel) {
-        purchaseOrderRepo.save(purchaseOrderModel);
-        return "your record is saved successfully !!";
+    @PostMapping
+    public String insertOrder(@RequestBody @Valid List<PurchaseOrderModel> purchaseOrderModel) {
+
+        List<PurchaseOrderModel> purchaseOrderModels= purchaseOrderRepo.saveAll(purchaseOrderModel);
+           if(purchaseOrderModels!=null)
+            return "your record is saved successfully !!";
+           else
+            return "{projectId},{orderNumber},{orderValidTillDate} required!!";
+
     }
 
-    @GetMapping("/getAllOrders")
-    public List<PurchaseOrderModel> getAllOrders() {
-        return (List<PurchaseOrderModel>) purchaseOrderRepo.findAll();
+    @GetMapping("/{projectId}")
+    public List<PurchaseOrderModel> getOrder(@PathVariable("projectId") int projectId){
+        return purchaseOrderService.getOrderBy(projectId);
     }
 
-    @GetMapping("/getOrderByProjectId/{projectId}")
-    public List<PurchaseOrderModel> getOrderBy(@PathVariable("projectId")  int projectId) {
-
-        return purchaseOrderRepo.findByProjectId(projectId);
+    @GetMapping("/{purchaseOrderId}")
+    public ResponseEntity<PurchaseOrderModel> getOrderById(@PathVariable("purchaseOrderId") int purchaseOrderId){
+        return purchaseOrderService.getOrderById(purchaseOrderId);
     }
 
-    @DeleteMapping("deleteByPurchaseOrderId/{purchaseOrderId}")
+
+    @DeleteMapping("/{purchaseOrderId}")
     public String deletePurchaseOrder(@PathVariable("purchaseOrderId") int purchaseOrderId){
-        PurchaseOrderModel purchaseOrderModel=purchaseOrderRepo.getOne(purchaseOrderId);
+       PurchaseOrderModel purchaseOrderModel=purchaseOrderRepo.findByPurchaseOrderId(purchaseOrderId);
 
-        purchaseOrderRepo.delete(purchaseOrderModel);
-        return "Your {purchaseOrderId} record is deleted successfully !!";
+        if(purchaseOrderModel!=null) {
+            purchaseOrderRepo.delete(purchaseOrderModel);
+            return "Your {purchaseOrderId} record is deleted successfully !!";
+        }else
+            return "Invalid {purchaseOrderId}";
     }
 
 
-    @PutMapping("/updatePurchaseOrder")
-    public List<PurchaseOrderModel> saveOrupdatePurchaseOrder(@RequestBody List<PurchaseOrderModel> purchaseOrderModel) {
-        purchaseOrderRepo.saveAll(purchaseOrderModel);
-        return purchaseOrderModel;
+    @PutMapping("/{purchaseOrderId}")
+    public ResponseEntity<PurchaseOrderModel> updateOrsave(@PathVariable("purchaseOrderId") int purchaseOrderId, @Valid @RequestBody PurchaseOrderModel purchaseOrderModel){
+        PurchaseOrderModel purchaseOrderModel1=purchaseOrderRepo.findByPurchaseOrderId(purchaseOrderId);
+        if(purchaseOrderModel1!=null) {
+            purchaseOrderModel1.setOrderDate(purchaseOrderModel.getOrderDate());
+            purchaseOrderModel1.setOrderNumber(purchaseOrderModel.getOrderNumber());
+            purchaseOrderModel1.setOrderValidTillDate(purchaseOrderModel.getOrderValidTillDate());
+            purchaseOrderModel1.setProjectId(purchaseOrderModel.getProjectId());
+            return new ResponseEntity<>(purchaseOrderRepo.save(purchaseOrderModel1), HttpStatus.OK);
+        } else
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @PutMapping("/updateByPurchaseOrderId/{purchaseOrderId}")
-    public PurchaseOrderModel updateOrsave(@PathVariable("purchaseOrderId") int purchaseOrderId,@RequestBody PurchaseOrderModel purchaseOrderModel){
+    @PutMapping("/{projectId}")
+    public ResponseEntity<PurchaseOrderModel> updateOrsaveBy(@PathVariable("projectId") int projectId, @Valid @RequestBody PurchaseOrderModel purchaseOrderModel){
+       PurchaseOrderModel purchaseOrderModel1=purchaseOrderRepo.findByprojectId(projectId);
 
-        return purchaseOrderRepo.save(purchaseOrderModel);
-    }
-
-    @PutMapping("/updateOrderByProjectId/{projectId}")
-    public PurchaseOrderModel updateOrsaveBy(@PathVariable("projectId") int projectId,@RequestBody PurchaseOrderModel purchaseOrderModel){
-
-        return purchaseOrderRepo.save(purchaseOrderModel);
+       if(purchaseOrderModel1!=null) {
+           purchaseOrderModel1.setOrderDate(purchaseOrderModel.getOrderDate());
+           purchaseOrderModel1.setOrderNumber(purchaseOrderModel.getOrderNumber());
+           purchaseOrderModel1.setOrderValidTillDate(purchaseOrderModel.getOrderValidTillDate());
+          // purchaseOrderModel1.setProjectId(purchaseOrderModel.getProjectId());
+           return new ResponseEntity<>(purchaseOrderRepo.save(purchaseOrderModel1), HttpStatus.OK);
+       }else
+             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
